prefs.js: fix Gtk.FileChooserNative input focus
The Gtk.FileChooserNative dialog didn't get input focus: all the
mouseclicks and keystrokes went to the prefs widget. To fix this we
need to set the transient-for property on the FileChooserNative to the
prefs widget. Since gtk_widget_get_toplevel is gone in Gtk4, implement
a get_toplevel function and use it in the clicked callback of the
button that launches the FileChooserNative dialog.